pyproject.toml
[project]
name = "obs-rec"
version = "0.1.0"
description = "Discord bot for automatic video recording and posting via OBS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.4.0",
    "obsws-python>=1.7.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pyright>=1.1.398",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = true
log_level = "INFO"
minversion = "6.0"
testpaths = "tests/"
asyncio_default_fixture_loop_scope = "function"

[tool.pyright]
pythonVersion = "3.12"
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "**/.venv",
  "tests/"
]
reportMissingTypeStubs = false
reportImplicitOverride = true
reportPrivateUsage = "warning"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
include = ["*.py"]
exclude = [
    ".git",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E203", "E402", "E501", "F811", "E741", "E266", "F401"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
